seperate temp and controls
The AC/Heat/Fan are in their own frame so they don't have access to the
Temp text field. Both frame buttons were sending a w=1. This was fine
when setting Heat/AC/Fan because temp was seen as empty. When setting
temp it had no access and no checks to set Heater AC Fan so the query
failed. No manual temp was set. Now the temp button sends a w=3 to
update the temp only.